Switch to using a branch based system...
The idea of having the dev and other experimental branches be hosted on
Github Pages for testing purposes is nice, but unnecessary in practice.
If we ever need to do that we can open up the test folder again.

For now, however, I’m switching to using branches for the dev (testing)
and main (prod) scenarios.