Added some cool tricks...
to get scrolling to work from other pages using a virtual id.

At some point, the onclick listener for the signup button can be done
away with (possibly in favor of another virtual id? maybe? Is this bad
practice?). Or we could just leave it alone.

Also, copy pasting code is the cause of all evils.